Refactor into modules.
There was a lot of functionality jammed into huge modules. These have been
broken out into smaller modules with some name changes to clear things up. This
is a reasonably massive breaking change, but needed to happen.

The `_ByteStringJson` prism has been removed for now.
