+++ src/Waargonaut/Types/JChar.hs
@@ -4,7 +4,6 @@
 {-# LANGUAGE FlexibleContexts       #-}
 {-# LANGUAGE FlexibleInstances      #-}
 {-# LANGUAGE FunctionalDependencies #-}
-{-# LANGUAGE LambdaCase             #-}
 {-# LANGUAGE MultiParamTypeClasses  #-}
 {-# LANGUAGE NoImplicitPrelude      #-}
 {-# LANGUAGE RankNTypes             #-}
@@ -13,20 +12,12 @@
 module Waargonaut.Types.JChar
   (
     -- * Types
-    HexDigit4 (..)
-  , HasHexDigit4 (..)
-  , JChar (..)
+    JChar (..)
   , AsJChar (..)
   , HasJChar (..)
-  , JCharEscaped (..)
-  , AsJCharEscaped (..)
-  , JCharUnescaped (..)
-  , AsJCharUnescaped (..)
 
     -- * Parser / Builder
   , parseJChar
-  , parseJCharEscaped
-  , parseJCharUnescaped
   , jCharBuilderWith
   , jCharBuilderTextL
   , jCharBuilderByteStringL
@@ -38,199 +29,66 @@ module Waargonaut.Types.JChar
   , charToJChar
   ) where
 
-import           Prelude                      (Char, Eq, Int, Ord, Show,
-                                               otherwise, quotRem, (&&), (*), (<), (>),
-                                               (+), (-), (/=), (<=), (==), (>=), (||))
+import           Prelude                          (Char, Eq, Ord, Show,
+                                                   otherwise, (/=))
 
-import           Control.Category             (id, (.))
-import           Control.Lens                 (Lens', Prism', Rewrapped,
-                                               Wrapped (..), failing, has, iso,
-                                               prism, prism', to, ( # ), (^?),
-                                               _Just)
+import           Control.Category                 (id, (.))
+import           Control.Lens                     (Lens', Prism', preview,
+                                                   prism, review, ( # ))
 
-import           Control.Applicative          (pure, (*>), (<$>), (<*>), (<|>))
-import           Control.Monad                ((=<<))
+import           Control.Applicative              ((<$>), (<|>))
 
-import           Data.Bits                    ((.&.))
-import           Data.Char                    (chr, ord)
-import           Data.Either                  (Either (..))
-import           Data.Foldable                (Foldable, any, asum, foldMap,
-                                               foldl)
-import           Data.Function                (const, ($))
-import           Data.Functor                 (Functor)
-import           Data.Maybe                   (Maybe (..), fromMaybe)
-import           Data.Monoid                  (Monoid)
-import           Data.Semigroup               (Semigroup, (<>))
-import           Data.Traversable             (Traversable, traverse)
+import           Data.Bits                        ((.&.))
+import           Data.Char                        (ord)
+import           Data.Either                      (Either (..))
+import           Data.Foldable                    (Foldable, asum, foldMap)
+import           Data.Function                    (($))
+import           Data.Functor                     (Functor)
+import           Data.Maybe                       (Maybe (..), fromMaybe)
+import           Data.Monoid                      (Monoid)
+import           Data.Semigroup                   (Semigroup, (<>))
+import           Data.Traversable                 (Traversable)
 
-import qualified Data.Text.Internal           as Text
+import qualified Data.Text.Internal               as Text
 
-import           Data.Digit                   (HeXDigit, HeXaDeCiMaL)
-import qualified Data.Digit                   as D
+import           Data.Digit                       (HeXDigit, HeXaDeCiMaL)
+import qualified Data.Digit                       as D
 
-import           Data.Text.Lazy.Builder       (Builder)
-import qualified Data.Text.Lazy.Builder       as TB
+import           Data.Text.Lazy.Builder           (Builder)
+import qualified Data.Text.Lazy.Builder           as TB
 
-import qualified Data.ByteString.Lazy.Builder as BB
+import qualified Data.ByteString.Lazy.Builder     as BB
 
-import           Waargonaut.Types.Whitespace  (Whitespace (..),
-                                               escapedWhitespaceChar,
-                                               unescapedWhitespaceChar,
-                                               _WhitespaceChar)
+import           Waargonaut.Types.Whitespace      (unescapedWhitespaceChar)
 
-import           Text.Parser.Char             (CharParsing, char, satisfy)
+import           Text.Parser.Char                 (CharParsing)
+
+import           Waargonaut.Types.JChar.HexDigit4 (HexDigit4 (..))
+
+import           Waargonaut.Types.JChar.Escaped   (AsEscaped (..), Escaped (..),
+                                                   charToEscaped, escapedToChar,
+                                                   parseEscaped)
+
+import           Waargonaut.Types.JChar.Unescaped (AsUnescaped (..), Unescaped,
+                                                   parseUnescaped)
 
 -- $setup
 -- >>> :set -XOverloadedStrings
--- >>> import Control.Monad (return)
+-- >>> import Data.Function (($))
 -- >>> import Data.Either(Either (..), isLeft)
 -- >>> import Data.Digit (HeXDigit(..))
--- >>> import qualified Data.Digit as D
--- >>> import Waargonaut.Decode.Error (DecodeError)
 -- >>> import Utils
+-- >>> import Waargonaut.Decode.Error (DecodeError)
+-- >>> import Waargonaut.Types.Whitespace
+-- >>> import Waargonaut.Types.JChar.Unescaped
+-- >>> import Waargonaut.Types.JChar.Escaped
+-- >>> import Waargonaut.Types.JChar
 ----
-
--- | JSON Characters may be single escaped UTF16 "\uab34".
-data HexDigit4 d =
-  HexDigit4 d d d d
-  deriving (Eq, Show, Ord, Functor, Foldable, Traversable)
-
--- | Typeclass for things that contain a 'HexDigit4'.
-class HasHexDigit4 c d | c -> d where
-  hexDigit4 :: Lens' c (HexDigit4 d)
-
-instance HasHexDigit4 (HexDigit4 d) d where
-  hexDigit4 = id
-
--- | Type to specify that this character is unescaped and may be represented
--- using a normal Haskell 'Char'.
-newtype JCharUnescaped =
-  JCharUnescaped Char
-  deriving (Eq, Ord, Show)
-
-instance JCharUnescaped ~ t => Rewrapped JCharUnescaped t
-instance Wrapped JCharUnescaped where
-  type Unwrapped JCharUnescaped = Char
-  _Wrapped' = iso (\ (JCharUnescaped x) -> x) JCharUnescaped
-
--- | Typeclass for things that may used as an unescaped JChar.
-class AsJCharUnescaped a where
-  _JCharUnescaped :: Prism' a JCharUnescaped
-
-instance AsJCharUnescaped JCharUnescaped where
-  _JCharUnescaped = id
-
-instance AsJCharUnescaped Char where
-  _JCharUnescaped = prism'
-    (\(JCharUnescaped c) -> c)
-    (\c ->  if any ($ c) excluded then Nothing
-            else Just (JCharUnescaped c)
-    )
-    where
-      excluded =
-        [ (== '\NUL')
-        , (== '"')
-        , (== '\\')
-        , \x ->
-            let
-              c = ord x
-            in
-              (c < 0x20 && c > 0x21) ||  -- "%x20-21"
-              (c < 0x23 && c > 0x5B) ||  -- "%x23-5B"
-              (c < 0x5D && c > 0x10FFFF) -- "%x5D-10FFFF"
-        ]
-
--- | Things that may be escaped in a JSON string.
-data JCharEscaped digit
-  = QuotationMark
-  | ReverseSolidus
-  | Solidus
-  | Backspace
-  | WhiteSpace Whitespace
-  | Hex ( HexDigit4 digit )
-  deriving (Eq, Ord, Show, Functor, Foldable, Traversable)
-
--- | Typeclass for things that may be used as an escaped JChar.
-class AsJCharEscaped r digit | r -> digit where
-  _JCharEscaped   :: Prism' r (JCharEscaped digit)
-  _QuotationMark  :: Prism' r ()
-  _ReverseSolidus :: Prism' r ()
-  _Solidus        :: Prism' r ()
-  _Backspace      :: Prism' r ()
-  _WhiteSpace     :: Prism' r Whitespace
-  _Hex            :: Prism' r (HexDigit4 digit)
-
-  _QuotationMark  = _JCharEscaped . _QuotationMark
-  _ReverseSolidus = _JCharEscaped . _ReverseSolidus
-  _Solidus        = _JCharEscaped . _Solidus
-  _Backspace      = _JCharEscaped . _Backspace
-  _WhiteSpace     = _JCharEscaped . _WhiteSpace
-  _Hex            = _JCharEscaped . _Hex
-
-instance AsJCharEscaped (JCharEscaped digit) digit where
-  _JCharEscaped = id
-  _QuotationMark = prism (const QuotationMark)
-    (\ x -> case x of
-        QuotationMark -> Right ()
-        _             -> Left x
-    )
-  _ReverseSolidus = prism (const ReverseSolidus)
-    (\ x -> case x of
-        ReverseSolidus -> Right ()
-        _              -> Left x
-    )
-  _Solidus = prism (const Solidus)
-    (\ x -> case x of
-        Solidus -> Right ()
-        _       -> Left x
-    )
-  _Backspace = prism (const Backspace)
-    (\ x -> case x of
-        Backspace -> Right ()
-        _         -> Left x
-    )
-  _WhiteSpace = prism WhiteSpace
-    (\ x -> case x of
-        WhiteSpace y1 -> Right y1
-        _             -> Left x
-    )
-  _Hex = prism Hex
-    (\ x -> case x of
-        Hex y1 -> Right y1
-        _      -> Left x
-    )
-
--- | Convert a given 'HexDigit4' to a Haskell 'Char'.
-hexDigit4ToChar
-  :: HeXaDeCiMaL digit
-  => HexDigit4 digit
-  -> Char
-hexDigit4ToChar (HexDigit4 a b c d) =
-  chr (foldl (\acc x -> 16 * acc + (D.integralHexadecimal # x)) 0 [a,b,c,d])
-
-charToHexDigit4
-  :: Char
-  -> Maybe (HexDigit4 HeXDigit)
-charToHexDigit4 x = if x >= '\x0' && x <= '\xffff'
-  then mkHexDigit4 =<< traverse (^? D.integralHexadecimal) (collectHexValues 4 [] (getRemainder (ord x)))
-  else Nothing
-  where
-    mkHexDigit4 (a:b:c:d:_) = Just (HexDigit4 a b c d)
-    mkHexDigit4 _           = Nothing
-
-    getRemainder n = quotRem n 16
-
-    collectHexValues :: Int -> [Int] -> (Int,Int) -> [Int]
-    collectHexValues 0 acc _     = acc
-    collectHexValues n acc (0,0) = collectHexValues (n - 1) (0:acc) (0,0)
-    collectHexValues n acc (q,r) = collectHexValues (n - 1) (r:acc) (getRemainder q)
-    {-# INLINE collectHexValues #-}
-{-# INLINE charToHexDigit4 #-}
 
 -- | A JChar may be unescaped or escaped.
 data JChar digit
-  = EscapedJChar ( JCharEscaped digit )
-  | UnescapedJChar JCharUnescaped
+  = EscapedJChar ( Escaped digit )
+  | UnescapedJChar Unescaped
   deriving (Eq, Ord, Show, Functor, Foldable, Traversable)
 
 -- | Typeclass for things that have a 'JChar'.
@@ -243,8 +101,8 @@ instance HasJChar (JChar digit) digit where
 -- | Typeclass for things that be used as a 'JChar'.
 class AsJChar r digit | r -> digit where
   _JChar          :: Prism' r (JChar digit)
-  _EscapedJChar   :: Prism' r (JCharEscaped digit)
-  _UnescapedJChar :: Prism' r JCharUnescaped
+  _EscapedJChar   :: Prism' r (Escaped digit)
+  _UnescapedJChar :: Prism' r Unescaped
 
   _EscapedJChar   = _JChar . _EscapedJChar
   _UnescapedJChar = _JChar . _UnescapedJChar
@@ -262,40 +120,23 @@ instance AsJChar (JChar digit) digit where
         _                 -> Left x
     )
 
-instance AsJCharEscaped (JChar digit) digit where
-  _JCharEscaped = _JChar . _JCharEscaped
+instance AsEscaped (JChar digit) digit where
+  _Escaped = _JChar . _Escaped
 
-instance AsJCharUnescaped (JChar digit) where
-  _JCharUnescaped = _JChar . _JCharUnescaped
+instance AsUnescaped (JChar digit) where
+  _Unescaped = _JChar . _Unescaped
 
 -- instance AsJChar Char HeXDigit where
 -- Don't implement this, it's not a lawful prism.
 
-escapedJCharToChar :: JCharEscaped HeXDigit -> Char
-escapedJCharToChar = \case
-  QuotationMark  -> '"'
-  ReverseSolidus -> '\\'
-  Solidus        -> '/'
-  Backspace      -> '\b'
-  WhiteSpace wc  -> unescapedWhitespaceChar wc
-  Hex hd         -> hexDigit4ToChar hd
-
-charToEscapedJChar :: Char -> Maybe (JCharEscaped HeXDigit)
-charToEscapedJChar c = case c of
-  '"'  -> Just QuotationMark
-  '\\' -> Just ReverseSolidus
-  '/'  -> Just Solidus
-  '\b' -> Just Backspace
-  _    -> c ^? failing (_WhitespaceChar . to WhiteSpace) (to charToHexDigit4 . _Just . to Hex)
-
 jCharToChar :: JChar HeXDigit -> Char
-jCharToChar (UnescapedJChar uejc) = _JCharUnescaped # uejc
-jCharToChar (EscapedJChar ejc)    = escapedJCharToChar ejc
+jCharToChar (UnescapedJChar uejc) = review _Unescaped uejc
+jCharToChar (EscapedJChar ejc)    = escapedToChar ejc
 
 charToJChar :: Char -> Maybe (JChar HeXDigit)
-charToJChar c = -- Order matters... oh dear.
-  (UnescapedJChar <$> c ^?_JCharUnescaped) <|>
-  (EscapedJChar <$> charToEscapedJChar c)
+charToJChar c =
+  (UnescapedJChar <$> preview _Unescaped c) <|>
+  (EscapedJChar <$> charToEscaped c)
 
 utf8SafeChar :: Char -> Maybe Char
 utf8SafeChar c | ord c .&. 0x1ff800 /= 0xd800 = Just c
@@ -317,107 +158,6 @@ jCharToUtf8Char :: JChar HeXDigit -> Maybe Char
 jCharToUtf8Char = utf8SafeChar . jCharToChar
 {-# INLINE jCharToUtf8Char #-}
 
--- | Parse a single 'HexDigit4'.
---
--- >>> testparse parseHexDigit4 "1234" :: Either DecodeError (HexDigit4 HeXDigit)
--- Right (HexDigit4 HeXDigit1 HeXDigit2 HeXDigit3 HeXDigit4)
---
--- >>> testparse parseHexDigit4 "12aF" :: Either DecodeError (HexDigit4 HeXDigit)
--- Right (HexDigit4 HeXDigit1 HeXDigit2 HeXDigita HeXDigitF)
---
--- >>> testparse parseHexDigit4 "aBcD" :: Either DecodeError (HexDigit4 HeXDigit)
--- Right (HexDigit4 HeXDigita HeXDigitB HeXDigitc HeXDigitD)
---
--- >>> testparsetheneof parseHexDigit4 "12aF" :: Either DecodeError (HexDigit4 HeXDigit)
--- Right (HexDigit4 HeXDigit1 HeXDigit2 HeXDigita HeXDigitF)
---
--- >>> testparsethennoteof parseHexDigit4 "12aFx" :: Either DecodeError (HexDigit4 HeXDigit)
--- Right (HexDigit4 HeXDigit1 HeXDigit2 HeXDigita HeXDigitF)
-parseHexDigit4 ::
-  ( CharParsing f, HeXaDeCiMaL digit ) =>
-  f ( HexDigit4 digit )
-parseHexDigit4 = HexDigit4
-  <$> D.parseHeXaDeCiMaL
-  <*> D.parseHeXaDeCiMaL
-  <*> D.parseHeXaDeCiMaL
-  <*> D.parseHeXaDeCiMaL
-
--- | Parse an unescaped JSON character.
---
--- >>> testparse parseJCharUnescaped "a"
--- Right (JCharUnescaped 'a')
---
--- >>> testparse parseJCharUnescaped "\8728"
--- Right (JCharUnescaped '\8728')
---
--- >>> testparsetheneof parseJCharUnescaped "a"
--- Right (JCharUnescaped 'a')
---
--- >>> testparsethennoteof parseJCharUnescaped "ax"
--- Right (JCharUnescaped 'a')
-parseJCharUnescaped ::
-  CharParsing f =>
-  f JCharUnescaped
-parseJCharUnescaped =
-  JCharUnescaped <$> satisfy (has _JCharUnescaped)
-
--- | Parse an escapted JSON character.
---
--- >>> testparse parseJCharEscaped "\\\""
--- Right QuotationMark
---
--- >>> testparse parseJCharEscaped "\\\\"
--- Right ReverseSolidus
---
--- >>> testparse parseJCharEscaped "\\/"
--- Right Solidus
---
--- >>> testparse parseJCharEscaped "\\b"
--- Right Backspace
---
--- >>> testparse parseJCharEscaped "\\f"
--- Right (WhiteSpace LineFeed)
---
--- >>> testparse parseJCharEscaped "\\n"
--- Right (WhiteSpace NewLine)
---
--- >>> testparse parseJCharEscaped "\\r"
--- Right (WhiteSpace CarriageReturn)
---
--- >>> testparse parseJCharEscaped "\\t"
--- Right (WhiteSpace HorizontalTab)
---
--- >>> testparse parseJCharEscaped "\\u1234" :: Either DecodeError (JCharEscaped HeXDigit)
--- Right (Hex (HexDigit4 HeXDigit1 HeXDigit2 HeXDigit3 HeXDigit4))
---
--- >>> testparsetheneof parseJCharEscaped "\\t"
--- Right (WhiteSpace HorizontalTab)
---
--- >>> testparsethennoteof parseJCharEscaped "\\tx"
--- Right (WhiteSpace HorizontalTab)
-parseJCharEscaped ::
-  (CharParsing f, HeXaDeCiMaL digit) =>
-  f ( JCharEscaped digit )
-parseJCharEscaped =
-  let
-    z =
-      asum ((\(c, p) -> char c *> pure p) <$>
-        [
-          ('"' , QuotationMark)
-        , ('\\', ReverseSolidus)
-        , ('/' , Solidus)
-        , ('b' , Backspace)
-        , (' ' , WhiteSpace Space)
-        , ('f' , WhiteSpace LineFeed)
-        , ('n' , WhiteSpace NewLine)
-        , ('r' , WhiteSpace CarriageReturn)
-        , ('t' , WhiteSpace HorizontalTab)
-        ])
-    h =
-      Hex <$> (char 'u' *> parseHexDigit4)
-  in
-    char '\\' *> (z <|> h)
-
 -- | Parse a JSON character.
 --
 -- >>> testparse parseJChar "\\u1234" :: Either DecodeError (JChar HeXDigit)
@@ -430,31 +170,17 @@ parseJCharEscaped =
 -- Right (EscapedJChar (WhiteSpace CarriageReturn))
 --
 -- >>> testparsetheneof parseJChar "a"
--- Right (UnescapedJChar (JCharUnescaped 'a'))
+-- Right (UnescapedJChar (Unescaped 'a'))
 --
 -- >>> testparsethennoteof parseJChar "ax"
--- Right (UnescapedJChar (JCharUnescaped 'a'))
+-- Right (UnescapedJChar (Unescaped 'a'))
 parseJChar ::
   (CharParsing f, HeXaDeCiMaL digit) =>
   f ( JChar digit )
 parseJChar = asum
-  [ EscapedJChar <$> parseJCharEscaped
-  , UnescapedJChar <$> parseJCharUnescaped
+  [ EscapedJChar <$> parseEscaped
+  , UnescapedJChar <$> parseUnescaped
   ]
-
--- | Convert a 'JChar' to a Haskell 'Char'.
--- jCharToChar
---   :: HeXaDeCiMaL digit
---   => JChar digit
---   -> Char
--- jCharToChar (UnescapedJChar (JCharUnescaped c)) = c
--- jCharToChar (EscapedJChar jca) = case jca of
---     QuotationMark   -> '"'
---     ReverseSolidus  -> '\\'
---     Solidus         -> '/'
---     Backspace       -> '\b'
---     (WhiteSpace ws) -> _WhiteSpace # ws
---     Hex hexDig4     -> hexDigit4ToChar hexDig4
 
 -- | Using the given function, return the builder for a single 'JChar'.
 jCharBuilderWith
@@ -465,7 +191,7 @@ jCharBuilderWith
   => (Char -> builder)
   -> JChar digit
   -> builder
-jCharBuilderWith f (UnescapedJChar (JCharUnescaped c)) = f c
+jCharBuilderWith f (UnescapedJChar c) = f (review _Unescaped c)
 jCharBuilderWith f (EscapedJChar jca) = f '\\' <> case jca of
     QuotationMark           -> f '"'
     ReverseSolidus          -> f '\\'
